+++ camlib.py
@@ -1132,9 +1132,12 @@ class Geometry(object):
 
         merged_lines = linemerge(geos_lines)
         geos = geos_polys
-        for l in merged_lines:
-            geos.append(l)
-
+        try:
+            for l in merged_lines:
+                geos.append(l)
+        except TypeError:
+            geos.append(merged_lines)
+            
         # Add to object
         if self.solid_geometry is None:
             self.solid_geometry = []